PMM-6530 refactor changes
