+++ pmm-app/src/shared/components/Elements/Icons/Icons.styles.ts
@@ -5,7 +5,6 @@ import { GrafanaTheme } from '@grafana/data';
 
 export const getStyles = stylesFactory((theme: GrafanaTheme) => {
   const iconsFill = selectThemeVariant(
-    // @ts-ignore
     { light: 'black', dark: 'rgba(255, 255, 255, 0.8)' },
     theme.type,
   );
